tpm: Restructure tpm20_extend to use buffer and take hash as parameter
Restructure the tpm20_extend function to use a buffer for the
command to send to the TPM. The size of the buffer is calculated
from the size of tpm2_req_extend structure and the appended SHA1
hash.

Add the hash algorithm that's being used as a parameter to this
function.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>
Message-Id: <1469546393-6944-3-git-send-email-stefanb@linux.vnet.ibm.com>