tpm: Rework tpm_build_and_send_cmd() into tpm_simple_cmd()
Instead of passing an arbitrary buffer as a parameter to
tpm_build_and_send_cmd(), just support the simpler case of a parameter
that is only 0, 1, or 2 bytes in length.  This allows the callers to
pass the parameter directly and not worry about the big-endian
encoding.  Rename the function to tpm_simple_cmd().

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>
Message-Id: <1470928683-17480-7-git-send-email-kevin@koconnor.net>