error out if present cpus count changed during SMP bringup
if during SMP bringup a cpu is hotplugged, Seabios might
silently hung in

  while (cmos_smp_count != CountCPUs)

loop as cmos_smp_count might be less then CountCPUs if
SIPI were delivered to the hotplugged CPU.

Warn user about it and ask for reboot.

While at it rename CountCPUs to BroughtUpCPUs to make clear
what it is counting.

Signed-off-by: Igor Mammedov <imammedo@redhat.com>
Message-Id: <1474026858-47776-4-git-send-email-imammedo@redhat.com>