vgabios: Reorder video modes to work around a Windows bug
Windows Server 2016 and Windows 10 RS1 come with a bug in its blue screen
of death rendering logic which prevents it from generating crash dumps.

The bug does not manifest if Windows sees a suitable 32 bpp video mode
before a suitable 24 bpp video mode in the list of modes returned from
vgabios. This commit moves all 32 bpp modes to the front of the list to
make sure that this is always the case.

Signed-off-by: Ladi Prosek <lprosek@redhat.com>
Message-Id: <1476719584-5299-1-git-send-email-lprosek@redhat.com>