QEMU DMA: Add ability to write back to fw_cfg file
This adds a new command to the DMA protocol.  The command allows the
memory allocation of a fw_cfg file by BIOS and subsequent return of
the allocated address to QEMU.  The initial use case is for Windows VM
Generation ID, where QEMU needs to change the contents of fw_cfg
data at runtime, while still having BIOS allocate and manage the memory.

Signed-off-by: Ben Warren <ben@skyportsystems.com>
Message-Id: <9e7d92085a05674917baaacc96bd103acb70ac43.1484893132.git.ben@skyportsystems.com>