SeaVGABios/cbvga: Assume VGA compatible GPU in text-mode
Assume the GPU is VGA compatible when running in text-mode.
Advertise VGA modes.
Microsoft Windows relies on graphics mode 12 as last resort when
running in text mode. It is used even when not advertised.

Allows to boot Windows 7 in "Safe Mode" and shows the boot splash.

While the stdvga driver could be used instead, implementing it properly
in the cbvga driver allows to switch between text and gfx mode using
NVRAM option.

Signed-off-by: Patrick Rudolph <siro@das-labor.org>
Message-Id: <20170529172514.8308-3-siro@das-labor.org>