scsi: ensure LUN0 is added first
The introduction of REPORT LUNS caused a potential regression when
many disks are present on the same SCSI target.  If LUN0 is processed
too late, SeaBIOS might run out of memory and fail to register it.

There are two kinds of problems:

- QEMU presents the results of REPORT LUNS in the opposite order
  to the "-device" command-line options.  It's likely that LUN 0
  will be first in the command-line, and thus will be processed
  last by SeaBIOS.

- QEMU registers all LUNs in each target first.  It's thus possible
  that LUN 0 of target 1 is processed after many LUNs of target 0
  and not be able to allocate memory.

To fix both, the SCSI scans are modified to operate in two phases;
LUN 0 is scanned before the REPORT LUNS command is sent.  A possible
future enhancement is to remember the results of the scan in a
32-byte bitmap, and skip the REPORT LUNS command during the second
phase.

When multiple SCSI controllers are found, it's still possible to
have a regression because they are scanned in parallel.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Message-Id: <20170619162110.28156-2-pbonzini@redhat.com>