virtio-scsi: speed up SCSI bus scan
The introduction of REPORT LUNS caused virtio-scsi to send two SCSI
commands (INQUIRY and REPORT LUNS) to each target rather than just
one.

However, INQUIRY will only fail if _no_ LUN is present on that target.
If all devices have LUN>0 is present, INQUIRY will report a dummy
device.  Therefore, the LUN0 scan can be used to prepare a compact
bitmap of targets that we have to scan, and omit sending REPORT LUNS to
the others.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Message-Id: <20170619162110.28156-3-pbonzini@redhat.com>