seabios: build RSDT from XSDT
Old operating systems would like to have a rev1 (ACPI 1.0) FADT, but
new operating systems would like to have rev3 (ACPI 2.0).

Since old operating systems do not know about XSDTs, the
solution is to point the RSDT to a rev1 FADT and the XSDT to a
rev3 FADT.

But, edk2 is not able to handle QEMU providing two FADTs and barfs when
it sees the second; edk2 subscribes to the view that the platform code
(meaning not only OVMF, but transitively QEMU's ACPI table builder)
should not handle such hacks; it's common edk2 code that should handle
FADT rev1 vs. rev3 and RSDT vs. XSDT.

These patches make SeaBIOS follow the same model as edk2, the only
difference being how the two identify ACPI tables from the BIOS
linker/loader script.  For SeaBIOS, this task is actually much
simpler since it can just look into the RSDP: if QEMU only
provides an XSDT, SeaBIOS takes care of building the RSDT and
rev1 FADT to satisfy ACPI 1.0-compliant operating systems.

This part makes SeaBIOS build an RSDT out of an existing XSDT,
patching the RSDP to point to the RSDT.

Reviewed-by: Phil Dennis-Jordan <phil@philjordan.eu>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Message-Id: <20170726094235.14267-2-pbonzini@redhat.com>