build: allow setting tools path from the environment
Allow setting the path to as, ld, objcopy, objdump, strip and python
from the environment.

This is required for building SeaBIOS on FreeBSD, which will switch
the default ld to lld very soon, and lld is not capable of building
SeaBIOS at the moment.

Building SeaBIOS on FreeBSD after the switch to lld will require
setting LD=/path/to/gnu/ld at build time.

Signed-off-by: Roger Pau Monné <roger.pau@citrix.com>
Message-Id: <20180123153416.75513-1-roger.pau@citrix.com>