Fixes to the floppy driver, so it works on real (not emulated) hardware
Several fixes to the floppy driver, so it works on a real floppy controller
with a real floppy. Tested with Coreboot on a motherboard with an ITE IT8712
Super I/O chip (and its built-in floppy controller) and a 1.44MB floppy.

Only one floppy is supported for now and only 1.44MB, but more formats will be
added in the future.

Signed-off-by: Nikolay Nikolov <nickysn@users.sourceforge.net>
Message-Id: <20180130034147.6058-1-nickysn@users.sourceforge.net>