malloc: Fallback to low-zone when alloc from high-zone fails
When 32-bit code needs to allocate memory which remains permanent
after completion of the initialization phase, it doesn't care if
it is allocated from high-zone or low-zone.

Currently all these cases just attempt to allocate from high-zone
as low-zone memory is scarce.

However, in cases we don't have enough space left in high-zone,
it makes sense to use low-zone as a fallback.

This commit fixes an issue of a VM which failed to boot when
configured with 16 PVSCSI disk controllers. The VM failed to boot
because pvscsi_init_rings() failed to allocate ring_dsc when
high-zone didn't had enough space for it.

Reviewed-by: Mark Kanda <mark.kanda@oracle.com>
Signed-off-by: Liran Alon <liran.alon@oracle.com>
Message-Id: <20181113155341.133016-3-liran.alon@oracle.com>