SMBIOS: Add SMBIOS Type 6 Memory Module Information
Add support for obsolete SMBIOS Type 6 which describes the speed, type,
size and error status of each system memory module.

This is required by some guests to boot successfully.

Such an example is Cisco NGFW appliance which has a script which
runs every boot that parses this SMBIOS Type 6 information and if
it doesn't exists, it just fails to boot with an error of
"Unable to parse dmidecode. Restarting now!".

Reviewed-by: Liran Alon <liran.alon@oracle.com>
Reviewed-by: Ross Philipson <ross.philipson@oracle.com>
Signed-off-by: Arbel Moshe <arbel.moshe@oracle.com>
Message-Id: <20181128155422.29113-1-liran.alon@oracle.com>