Revert "mptable: Don't describe pci-to-pci bridges."
This reverts commit
3c3a3fa6522f ("mptable: Don't describe pci-to-pci bridges.”)

The reverted commit removed the description of non-root PCI busses
from the MPTable in claim they are not necessary.

However, it seems that some guests rely on this information in order to
correclty configure IOAPIC redirection-table entries for interrupts
originated from PCI devices on non-root PCI busses.

One such guest is "Extreme Networks OS".
We observed that if this guest is setup with an E1000 NIC behind a PCI
bridge, the OS wouldn't configure the IOAPIC redirection-table entry for
the IRQ that the E1000 PIN is connected to. Therefore, interrupts from
the NIC were not delivered to OS which caused guest to not have network
connectivity.

Fixes: 3c3a3fa6522f ("mptable: Don't describe pci-to-pci bridges.”)

Reviewed-by: Liran Alon <liran.alon@oracle.com>
Reviewed-by: Mark Kanda <mark.kanda@oracle.com>
Signed-off-by: Arbel Moshe <arbel.moshe@oracle.com>
Message-Id: <20181128164951.95230-1-liran.alon@oracle.com>