bios_date: Make bios date appear once in the ROM
We find all the places which cause BIOS_DATE to appear in the data
section and remove them - leaving the only place to contain the contents
of BIOS_DATE to be BiosDate at FSEG(0xfff5).

We do this by removing the usages of BIOS_DATE and by changing
strlen(BIOS_DATE) to sizeof(BIOS_DATE) which works at the preprocessing
step - removing the need to add BIOS_DATE as a string in the data
section.

We will use the fact that BiosDate is now the only place in FSEG
containing the date in the next commit.

Reviewed-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>
Reviewed-by: Arbel Moshe <arbel.moshe@oracle.com>
Signed-off-by: Sam Eiderman <shmuel.eiderman@oracle.com>
Signed-off-by: Liran Alon <liran.alon@oracle.com>
Message-Id: <20190517205726.34703-3-shmuel.eiderman@oracle.com>