csm: Fix boot priority translation
For CSM, the highest priority is zero. In SeaBIOS that means "don't", and
the highest priority is 1.

So we end up with the fun outcome that booting from NVMe worked only
when it *wasn't* selected as the primary boot target, because we don't
actually run the nvme_controller_setup() thread for an NVMe controller
if its boot prio is zero.

Signed-off-by: David Woodhouse <dwmw2@infradead.org>
Message-Id: <5948b65e805e5644b41172a1439c579866461897.camel@infradead.org>