virtio: extend virtio queue size to 256
Some linux kernels has a performance flaw in virtio block device access.
On some frequent disk access patterns, e.g. 1M read, the kernel produces
more block requests than needed. This happens because of virtio seg_max
parameter set to 126 (virtqueue_size - 2) which limits the maximum block
request to 516096 (126 * 4096_PAGE_SIZE).

Setting seg_max > 126 fixes the issue, however, not all linux kernels
allow that. The old ones have a restriction virtqueue_size >= seg_max.
This restriction is hardcoded and the kernel crashes in case of violation.
The restriction is relaxed in the recent kernels. Windows kernels don't
have such a restriction.

To increse seg_max and not to break the restriction, one can increase the
virtqueue size to 256 and seg_max to 254. To do that, seabios support of
256 virtqueue size is needed.

Signed-off-by: Denis Plotnikov <dplotnikov@virtuozzo.com>
Message-Id: <20191001072500.18944-1-dplotnikov@virtuozzo.com>