hw/usb-hid: handle devices with illegal max packet size
Some USB keyboards report 9 or 10-byte max packet sizes, instead
of the 8-byte max specified by the USB HID spec. Handle this by
increasing the size of the keyevent struct to 10 bytes, zeroizing
it before use, and using the key array size of the usbkeyinfo
struct as loop bounds rather than that of the keyevent struct
(since the former will always be smaller, and within spec) to
prevent buffer overflow.

Test: built/boot on Google Pixel Slate, observe keyboard functional

Signed-off-by: Matt DeVillier <matt.devillier@gmail.com>
Message-Id: <CAFTm+6A2p5KFJ0Wgg9yL7AaTU9sDqLOU8ZSDR49smCjzo-dtfw@mail.gmail.com>