hw/usb-hid: handle devices with illegal max packet size
Some USB keyboards report 9 or 10-byte max packet sizes, instead
of the 8-byte max specified by the USB HID spec. Handle this by
increasing the size of the keyevent struct to 10 bytes, and using the key
array size of the usbkeyinfo struct as loop bounds rather than that of the
keyevent struct (since the former will always be smaller, and within spec).

Test: built/boot on Google Pixel Slate, observe keyboard functional

Signed-off-by: Matt DeVillier <matt.devillier@gmail.com>
Message-Id: <CAFTm+6C8_0FuKQPXarfwTXOPRMMzn5agjbVT_7qCLE_MsR7kbA@mail.gmail.com>