boot: add config option for fast device init
  Hi,

> > > > Add a config option for fast device init (i.e. skip non-bootable
> > > > devices).  This optimization breaks boot in some rare configurations,
> > > > specifically in case the boot loader needs access to secondary disks.
>
> Unfortunately, I missed the initial email due to seabios mailing list
> problems a few weeks ago.

Quote above is the commit message, actual patch below.

> What's the reason to add a config option for this setting?  We
> typically make these choices at runtime using settings from fwcfg.

Well, there is a fwcfg setting already.  Skipping non-bootable devices
happens only in case there is a HALT line in the fw_cfg bootorder file,
which translates to "-boot strict={on,off}" on the qemu command line.

Problem is that libvirt runs qemu with strict=on and provides no way to
request strict=off instead.  Fixing that is on the way, but it'll take
some time as it affects not only libvirt itself but also the management
apps using libvirt.

So this patch allows to handles the reported regressions with a compile
time switch for the time being.  If you don't like it, fine, we can
also carry the patch downstream for a while until libvirt & co are
fixed, it should only be needed temporarily after all.

I suspect we are not the only ones running into this though, so it might
be useful to have that upstream too ...

take care,
  Gerd

------------------------------ cut here -------------------------
Message-Id: <20210408100223.6u3oe7vr43n6d63p@sirius.home.kraxel.org>