nvme: look for NVMe namespace in boot rom file
This patch allows specifying an NVMe namespace ID in the boot ROM file
in the form /pci@0cf8/*@*/nvme/ns@*, where the 2nd component is the
PCI device and the last component is the namespace ID.

E.g. booting from namespace 2:

    boot order:
    1: /pci@i0cf8/pci4e58,0001@6/nvme/ns@2
    2: /rom@genroms/bootfail.bin
    3: HAL
    ...
    |3ffa9000| Searching bootorder for: /pci@i0cf8/*@6/nvme/ns@2
    |3ffa9000| NS 2 prio 1
    |3ffa9000| phys_alloc zone=0x3ffbfeb8 size=80 align=10 ret=3ffac6e0
    (detail=0x3ffac6b0)
    |3ffa9000| phys_alloc zone=0x3ffbfeb8 size=24 align=10 ret=3ffac690
    (detail=0x3ffac660)
    |3ffa9000| Registering bootable: NVMe NS 2: 19073 MiB (39062500 512-byte
    blocks + 0-byte metadata)
     (type:2 prio:1 data:f4050)
    |3ffa9000| NVMe initialization complete!
    \3ffa9000/ End thread

If we find a specific namespace we're looking for then we stop probing
all other namespaces as there's no point in doing so. If a namespace is
not specified in the boot ROM file then we default to the old way of
booting from NVMe.

As a future optimization we can avoid keeping information for all
namespaces and operate on a single namespace at at time. This way we
won't run out of memory even if the NVMe controller has the maximum
number of namespaces (2^32 - 1).

The corresponding QEMU patches haven't been upstreamed yet.

Signed-off-by: Thanos Makatos <thanos.makatos@nutanix.com>

--

Resending patch since it never reached the list, including a more detailed
commit message.

Message-Id: <20210422091936.76378-1-thanos.makatos@nutanix.com>