fw/coreboot.c: Use coreboot table to find cbfs
The "cbfs master header" cbfs file is considered a legacy feature in
coreboot and is planned for removal in the master branch. Since 2015
the coreboot tables have exported info about the active cbfs.

This change uses the cbfs information in the coreboot tables instead
of following the cbfs master header pointer in the bootblock to find
cbfs. This change makes it possible to access CBFS fmap regions that
don't feature a "cbfs master header", which is the case with Googles
VBOOT solution.

This breaks compatibility with very old coreboot build (build before
fb5d5b16 "2015-07-14, cbtable: describe boot media"). Keeping
backward compatibility with the "cbfs master header" would complicate
the code.

Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>

Message-Id: <20210430082826.394372-1-arthur@aheymans.xyz>