make BUILD_MIN_BIOSTABLE configurable
BUILD_MIN_BIOSTABLE reserves space in the f-segment.  Some data
structures -- for example disk drives known to seabios -- must be
stored there, so the space available here limits the number of
devices seabios is able to manage.

This patch adds an config option for BUILD_MIN_BIOSTABLE so the size of
the reservation can be configured at build time.  Default is 8k for bios
images being 256k or larger in size (32bit code is moved off then so we
have more room in the f-segment), 2k (current value) otherwise.

Signed-off-by: Gerd Hoffmann <kraxel@redhat.com>
Message-Id: <20210527085227.1029147-1-kraxel@redhat.com>