Add an option for debug output on the vga console
Only available if threads are disabled, to prevent an undiagnosed call16 panic.

A new vga_console_active() utility function is added, and debugging output is
only forwarded to the vga console if the vga console has been enabled.

Only debugging output from 32bit flat regions is emitted, as vga console output
is only supported from within such code. This is detected with a !MODESEGMENT
check inside the debugging output code, and a new ASSERT32FLAT() check within the
vga output code.

This change has been briefly tested on a kfsn4-dre and in qemu, using serial
debugging in the latter to verify that 16bit output appears over serial that
does not appear on the vga console.

Signed-off-by: Karl Semich <0xloem@gmail.com>
Message-Id: <20220312122333.33147-1-0xloem@gmail.com>