reset: force standard PCI configuration access
After a reset of a QEMU -machine q35 guest, the PCI Express
Enhanced Configuration Mechanism is disabled and the variable
mmconfig no longer matches the configuration register MCHBAR
of the Q35 chipset. Until the variable mmconfig is reset to 0,
all pci_config_*() functions no longer work.

The variable mmconfig is located in the read-only F-segment.
To reset it the pci_config_*() functions are needed, but they
do not work.

Replace all pci_config_*() calls with Standard PCI Configuration
Mechanism pci_ioconfig_*() calls until mmconfig is reset to 0.

This fixes

In resume (status=0)
In 32bit resume
Attempting a hard reboot
Unable to unlock ram - bridge not found

and a reset loop with QEMU -accel tcg.

Signed-off-by: Volker Rümelin <vr_qemu@t-online.de>
Reviewed-by: Gerd Hoffmann <kraxel@redhat.com>
Message-Id: <20220320094633.2239-2-vr_qemu@t-online.de>