MP: fix mptable interrupt source generation for pci devices
Set the correct IOAPIC INTIN pin number for pci bus interrupt
sources during MP spec table generation (on emulators). Currently,
the pin number is set to the interrupt line field in the device's
configuration space, which is set to either IRQ 10 or 11 as the
boot rom driver expects. This works since qemu maps all ISA
compatible PIC IRQs onto IOAPIC pins 0-15, including PIRQs. But it
will break if, for some reason, the IRQ is routed to something
other than the INTLINE value using, for eg. ACPI _CRS. This patch
ensures the pin number is set to the correct value (16-23) that the
INTx pin is routed to in APIC mode, in agreement with the ACPI _PRT
provided routing.

Tested on a Linux 5.17.2 guest on qemu with the pci=nomsi and
acpi=noirq boot parameters to force MP table parsing for interrupt
sources.

Signed-off-by: Jay Khandkar <jaykhandkar2002@gmail.com>
Message-Id: <20220415110334.11996-1-jaykhandkar2002@gmail.com>