romlayout.S: Add missing SHF_ALLOC flag to .fixedaddr.\addr
It does not make sense to reference a SHF_ALLOC section from a
non-SHF_ALLOC section via R_386_PC16. Conceptually, even if a
non-SHF_ALLOC is loaded as part of the memory image, the distance
between it and a SHF_ALLOC section may not be a constant, so the linker
cannot reasonably resolve the relocation.

GNU ld allows it but lld will warn. Add the SHF_ALLOC flag.

Signed-off-by: Fangrui Song <maskray@google.com>
Message-Id: <20220713013508.3428114-2-maskray@google.com>