add hwerr_printf function for threads
Printing to the screen from threads doesn't work, because that involves
a switch to real mode for using int10h services.

Add a string buffer and hwerr_printf() helper functions to store error
messages. Print the buffer later, after device initialization, from main
thread in case it is not empty.

Signed-off-by: Gerd Hoffmann <kraxel@redhat.com>
Message-Id: <20230503092201.487130-2-kraxel@redhat.com>