an issue with win10 boot and different compiler versions
On Sat, Feb 10, 2024 at 11:17:54PM +0300, Michael Tokarev wrote:
> So.. the difference is vgabios only, not seabios (vgabios-stdvga in this case).
>
> And I can't get it to work with debugging vgabios, it always fails even with DEBUG_LEVEL=2
> (and level-1 logging isn't useful).
>
> I was able to capture logs just for the non-working version, so there's nothing to
> compare it against.  So I tried a different machine type in qemu, the one which
> works, which uses SMBIOS 3.0 (q35-8.2).

Thanks for testing.  So, if I understand the issue correctly:
1. If smbios v3 is used then the problem does not occur.
2. If gcc v13 is used to compile vgabios then the problem does not occur.
3. If smbios v2 is used and gcc v12 is used then win10 can not boot.
Is that correct?

A strange issue.  Issues like this tend to be very difficult to track
down.

As a random guess, one possibility is that it could be related to
vgabios stack size usage.  You could try always enabling the "extra
vga stack" with a change like:

Message-Id: <ZdUAfhd_lKvlRUDX@morn>