Add basic MXM 3.0 interrupt support
VGAROMs on MXM graphics cards need certain int15h functions present
to function properly.

On HP EliteBook 8560w with coreboot and Quadro 2000M, this warning
is displayed for 30 seconds, making every boot extremely slow:

	ERROR: Valid MXM Structure not found
	POST halted for 30 seconds, P-state limited to P10...

This patch implements the minimum functionality to get rid of it: the
functions 0 and 1, version 3.0 of the specification [1]. Older version
2.1 is not implemented.

Functions are enabled only if romfile "mxm-30-sis" exists. These functions
are specific to the MXM revision, but not the board or GPU. Some boards
might require more of the optional functions to be implemented, however.

- Function 0 returns information about supported functions and revision.

- Function 1 returns a pointer to a MXM configuration structure in low
memory. This is copied from romfile "mxm-30-sis". It can be extracted
from vendor BIOS, by using this same interrupt. I wrote a tool [2] to do
that from Linux.

TEST: HP EliteBook 8560w boots without error and delay, graphics work.

[1]: MXM Graphics Module Software Specification Version 3.0, Revision 1.1
[2]: https://codeberg.org/Riku_V/mxmdump/

Signed-off-by: Riku Viitanen <riku.viitanen@protonmail.com>
Acked-by: Paul Menzel <pmenzel@molgen.mpg.de>
Message-Id: <GICJtUPTMGjC_Q3wqzoiVaZYPq-HCF7E3nRT18tECk1CpVNuPCnEb_EnaaRB2pJNIohuRnkwdyETKo8362j5eeYyMJLfZx1IlkdJPmLJdxA=@protonmail.com>